scripts/saml_script/requirements.txt

beautifulsoup4==4.12.0
boto3==1.26.97
botocore==1.29.97
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
cryptography==39.0.2
idna==3.4
jmespath==1.0.1
pycparser==2.21
pyspnego==0.8.0
python-dateutil==2.8.2
requests==2.28.2
requests-ntlm==1.2.0
s3transfer==0.6.0
six==1.16.0
soupsieve==2.4
urllib3==1.26.15
python-dotenv==1.0.1
pywin32==306
selenium==4.20.0
